+++ branches/mc.struct2/src/plugins/switch/federation/switch_federation.c
@@ -390,7 +390,7 @@ static char *adapter_name_check(char *network)
 }
 
 int switch_p_build_jobinfo(switch_jobinfo_t switch_job, char *nodelist, 
-			uint32_t *tasks_per_node, int cyclic_alloc, 
+			uint16_t *tasks_per_node, int cyclic_alloc, 
 			char *network) 
 {
 	hostlist_t list = NULL;
